+++ src/app/core/ui/pages/page-in-progress/page-in-progress.module.ts
@@ -1,9 +1,11 @@
 import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
 import { PageInProgressComponent } from './page-in-progress.component';
+import { RouterModule } from '@angular/router';
+import { TranslateModule } from '@ngx-translate/core';
 
 @NgModule({
   declarations: [PageInProgressComponent],
-  imports: [CommonModule],
+  imports: [CommonModule, RouterModule, TranslateModule],
 })
 export class PageInProgressModule {}