+++ CHANGELOG.md
@@ -31,3 +31,4 @@
 - Mise en place du footer
 - Mise en place du breadcrumb
 - Mise en place de la page not found
+- Mise en place de la page in progress